Fix architecture to only use Hickory instead of old dns system
