+++ hackflare_dns/src/dns/mod.rs
@@ -1,12 +1,5 @@
 pub mod config;
-pub mod engine;
-pub mod manager;
-pub mod records;
 pub mod recursive;
-pub mod registry;
-pub mod zone;
+pub mod wire;
 
 pub use config::DnsConfig;
-pub use manager::DnsManager;
-pub use records::Record;
-pub use zone::Zone;
