--- hackflare_dns/src/dns/zone.rs
@@ -1,8 +1,0 @@
-use crate::dns::Record;
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Zone {
-    pub name: String,
-    pub records: Vec<Record>,
-}
